refactor(interfaces):  rename pages into views
