+++ lib/core/entities/movies/movie.dart
@@ -1,16 +1,11 @@
 import 'package:equatable/equatable.dart';
 
-import 'genre.dart';
-import 'movie_collection.dart';
-import 'spoken_language.dart';
+import '../spoken_language.dart';
 
-class MovieDetail extends Equatable {
-  final bool adult;
+class Movie extends Equatable {
+  final bool isAdult;
   final String? backdropPath;
-  final MovieCollection? movieCollection;
-  final int budget;
-  final List<Genre> genres;
-  final String? homepage;
+  final List<int> genreIds;
   final int id;
 
   /// Its in `ISO 639 1` form. Convert it after fetch List of [SpokenLanguage]
@@ -21,22 +16,14 @@ class MovieDetail extends Equatable {
   final double popularity;
   final String? posterPath;
   final DateTime? releaseDate;
-  final int revenue;
-  final int runtime;
-  final String status;
-  final String tagline;
   final String title;
-  final bool video;
   final double voteAverage;
   final int voteCount;
 
-  const MovieDetail({
-    required this.adult,
+  const Movie({
+    required this.isAdult,
     this.backdropPath,
-    this.movieCollection,
-    required this.budget,
-    required this.genres,
-    this.homepage,
+    required this.genreIds,
     required this.id,
     required this.language,
     required this.originalTitle,
@@ -44,12 +31,7 @@ class MovieDetail extends Equatable {
     required this.popularity,
     this.posterPath,
     this.releaseDate,
-    required this.revenue,
-    required this.runtime,
-    required this.status,
-    required this.tagline,
     required this.title,
-    required this.video,
     required this.voteAverage,
     required this.voteCount,
   });
@@ -57,11 +39,9 @@ class MovieDetail extends Equatable {
   @override
   List<Object?> get props {
     return [
-      adult,
+      isAdult,
       backdropPath,
-      budget,
-      genres,
-      homepage,
+      genreIds,
       id,
       language,
       originalTitle,
@@ -69,12 +49,7 @@ class MovieDetail extends Equatable {
       popularity,
       posterPath,
       releaseDate,
-      revenue,
-      runtime,
-      status,
-      tagline,
       title,
-      video,
       voteAverage,
       voteCount,
     ];
