+++ lib/core/entities/movies/movie_collection.dart
