+++ lib/core/entities/movies/movie_detail.dart
@@ -1,11 +1,16 @@
 import 'package:equatable/equatable.dart';
 
-import 'spoken_language.dart';
+import '../genre.dart';
+import 'movie_collection.dart';
+import '../spoken_language.dart';
 
-class Movie extends Equatable {
-  final bool isAdult;
+class MovieDetail extends Equatable {
+  final bool adult;
   final String? backdropPath;
-  final List<int> genreIds;
+  final MovieCollection? movieCollection;
+  final int budget;
+  final List<Genre> genres;
+  final String? homepage;
   final int id;
 
   /// Its in `ISO 639 1` form. Convert it after fetch List of [SpokenLanguage]
@@ -16,14 +21,22 @@ class Movie extends Equatable {
   final double popularity;
   final String? posterPath;
   final DateTime? releaseDate;
+  final int revenue;
+  final int runtime;
+  final String status;
+  final String tagline;
   final String title;
+  final bool video;
   final double voteAverage;
   final int voteCount;
 
-  const Movie({
-    required this.isAdult,
+  const MovieDetail({
+    required this.adult,
     this.backdropPath,
-    required this.genreIds,
+    this.movieCollection,
+    required this.budget,
+    required this.genres,
+    this.homepage,
     required this.id,
     required this.language,
     required this.originalTitle,
@@ -31,7 +44,12 @@ class Movie extends Equatable {
     required this.popularity,
     this.posterPath,
     this.releaseDate,
+    required this.revenue,
+    required this.runtime,
+    required this.status,
+    required this.tagline,
     required this.title,
+    required this.video,
     required this.voteAverage,
     required this.voteCount,
   });
@@ -39,9 +57,11 @@ class Movie extends Equatable {
   @override
   List<Object?> get props {
     return [
-      isAdult,
+      adult,
       backdropPath,
-      genreIds,
+      budget,
+      genres,
+      homepage,
       id,
       language,
       originalTitle,
@@ -49,7 +69,12 @@ class Movie extends Equatable {
       popularity,
       posterPath,
       releaseDate,
+      revenue,
+      runtime,
+      status,
+      tagline,
       title,
+      video,
       voteAverage,
       voteCount,
     ];
