+++ lib/core/use_cases/movies/get_movie_detail.dart
@@ -1,4 +1,4 @@
-import 'package:screen_pal/core/entities/movie_detail.dart';
+import 'package:screen_pal/core/entities/movies/movie_detail.dart';
 import 'package:screen_pal/core/repo/movies_repo.dart';
 
 class GetMovieDetail {
