+++ lib/core/use_cases/movies/get_movie_recommendations.dart
@@ -1,4 +1,4 @@
-import 'package:screen_pal/core/entities/movie.dart';
+import 'package:screen_pal/core/entities/movies/movie.dart';
 import 'package:screen_pal/core/repo/movies_repo.dart';
 
 class GetMovieRecommendations {
