+++ lib/interfaces/providers/movies/movie_detail_provider.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import 'package:screen_pal/configs/service_locator/locator.dart' as service;
-import 'package:screen_pal/core/entities/movie_detail.dart';
+import 'package:screen_pal/core/entities/movies/movie_detail.dart';
 import 'package:screen_pal/core/use_cases/movies/get_movie_detail.dart';
 
 part 'movie_detail_provider.g.dart';
