+++ lib/interfaces/providers/movies/movie_list_providers.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import 'package:screen_pal/configs/service_locator/locator.dart' as service;
-import 'package:screen_pal/core/entities/movie.dart';
+import 'package:screen_pal/core/entities/movies/movie.dart';
 import 'package:screen_pal/core/use_cases/movies/get_movie_recommendations.dart';
 import 'package:screen_pal/core/use_cases/movies/get_now_playing_movies.dart';
 import 'package:screen_pal/core/use_cases/movies/get_popular_movies.dart';
