+++ lib/interfaces/router/router.dart
@@ -3,10 +3,10 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import 'package:screen_pal/interfaces/pages/home_layout.dart';
-import 'package:screen_pal/interfaces/pages/movies/movie_detail_page.dart';
-import 'package:screen_pal/interfaces/pages/movies/movies_home_page.dart';
-import 'package:screen_pal/interfaces/pages/settings_page.dart';
+import 'package:screen_pal/interfaces/views/home_view_layout.dart';
+import 'package:screen_pal/interfaces/views/movies/movie_detail_view.dart';
+import 'package:screen_pal/interfaces/views/movies/movies_home_view.dart';
+import 'package:screen_pal/interfaces/views/settings_view.dart';
 import 'package:screen_pal/interfaces/widgets/errors/not_found_route_page.dart';
 
 final router = GoRouter(
@@ -22,11 +22,11 @@ final router = GoRouter(
   },
   routes: [
     ShellRoute(
-      builder: (_, __, child) => HomeLayout(child: child),
+      builder: (_, __, child) => HomeViewLayout(child: child),
       routes: [
         GoRoute(
           path: '/movies',
-          builder: (_, __) => const MoviesHomePage(),
+          builder: (_, __) => const MoviesHomeView(),
         ),
         GoRoute(
           path: '/tv_shows',
@@ -40,14 +40,14 @@ final router = GoRouter(
     ),
     GoRoute(
       path: '/settings',
-      builder: (_, __) => const SettingsPage(),
+      builder: (_, __) => const SettingsView(),
     ),
     GoRoute(
       path: '/movies/movie-:id',
       builder: (context, router) {
         final movieId = int.parse(router.pathParameters['id'] ?? '0');
 
-        return MovieDetailPage(movieId: movieId);
+        return MovieDetailView(movieId: movieId);
       },
     ),
   ],
