+++ lib/interfaces/views/movies/movies_home_view.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'package:screen_pal/configs/utils/riverpod_async_value_handlers.dart';
-import 'package:screen_pal/core/entities/movie.dart';
+import 'package:screen_pal/core/entities/movies/movie.dart';
 import 'package:screen_pal/interfaces/providers/movies/movie_list_providers.dart';
 import 'package:screen_pal/interfaces/widgets/carousel/movies_carousel.dart';
 import 'package:screen_pal/interfaces/widgets/list_view/movie_horiz_list_view.dart';
 
-class MoviesHomePage extends StatelessWidget {
-  const MoviesHomePage({super.key});
+class MoviesHomeView extends StatelessWidget {
+  const MoviesHomeView({super.key});
 
   @override
   Widget build(BuildContext context) {
