+++ lib/interfaces/widgets/carousel/movies_carousel.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:smooth_page_indicator/smooth_page_indicator.dart';
 
-import 'package:screen_pal/core/entities/movie.dart';
+import 'package:screen_pal/core/entities/movies/movie.dart';
 import 'package:screen_pal/infrastructures/api/tmdb_dio.dart';
 import 'package:screen_pal/interfaces/providers/extras/extras_providers.dart';
 import 'package:screen_pal/interfaces/router/app_navigator.dart';
@@ -11,8 +11,10 @@ import 'package:screen_pal/interfaces/widgets/default_network_image.dart';
 
 // Tests Utils:
 
-/// To check if which layout are rendered.
+/// To check which layout are rendered.
 const _contentContainerKey = Key('content-container');
+
+// To check movie contents:
 
 const _imageKey = Key('image-widget');
 const _movieTitleKey = Key('movie-title');
